requirements.txt
asgiref==3.2.7
Django==3.0.5
django-crispy-forms==1.9.0
django-debugtools==1.8
django-environ==0.4.5
django-seed==0.2.2
django-webpack-loader==0.7.0
djangomix==1.1.3
djangorestframework==3.11.0
Faker==4.0.3
Pillow==7.1.1
psycopg2==2.8.5
psycopg2-binary==2.8.5
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
gunicorn==20.0.4
gevent